% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_info.R
\name{domains_info}
\alias{domains_info}
\title{domains_info}
\usage{
domains_info(
  domains,
  excludes = "google.com, facebook.com, youtube.com",
  history = FALSE
)
}
\arguments{
\item{domains}{A comma separated string of domain names}

\item{excludes}{A comma separated string of domain names to exclude from search}

\item{history}{If true uses the payed Whois API, if false uses the raw data}
}
\value{
A data.frame with info about the given domains
}
\description{
A function that returns Whois data about a given list of domain names
}
\examples{
 info <- domains_info(
   domains = "domain1.com, domain2.com, domain3.com",
   excludes="domain2.com",
   history=TRUE)

 info2 <- domains_info(
   domains = "domain1.com, domain2.com, domain3.com",
   excludes="domain2.com")
}
