% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_url_timeline_chart.R
\name{draw_url_timeline_chart}
\alias{draw_url_timeline_chart}
\title{draw_url_timeline_chart}
\usage{
draw_url_timeline_chart(output, top_coord_urls = NULL, top_url_id = 1)
}
\arguments{
\item{output}{the output list resulting by the get_coord_shares function}

\item{top_coord_urls}{the output of get_top_coord_urls. If NULL a new top_url is created (default=NULL)}

\item{top_url_id}{a number indicating which url to use to draw the chart}
}
\value{
A data frame (grouped_df) containing the posts with the highest engagement shared in a coordinated way by the highly coordinated entities, with a set of attributes
}
\description{
Given an id from 1 to 10, this function draws an interactive timeline depicting how a link was shared on Facebook by coordinated and non coordinated shares
}
\examples{
  # get the top ten posts containing URLs shared by each network component and by engagement
  df <- get_top_news(output, order_by = "engagement", top=10)

  # get the top ten posts containing URLs shared in a coordinated way, by engagement
  df <- get_top_news(output, order_by = "engagement", top=10)

}
