% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coord_interval.R
\name{estimate_coord_interval}
\alias{estimate_coord_interval}
\title{estimate_coord_interval}
\usage{
estimate_coord_interval(
  ct_shares.df = NULL,
  mongo_database,
  mongo_url = NULL,
  q = 0.1,
  p = 0.5,
  clean_urls = FALSE,
  mongo_cluster = TRUE,
  keep_ourl_only = FALSE
)
}
\arguments{
\item{ct_shares.df}{a dataframe containing the Crowdtangle shares after running get_ctshares.R}

\item{mongo_database}{string: the name of the MongoDB used to host the collections.}

\item{mongo_url}{string: address of the MongoDB server in standard URI Format. Set to NULL to avoid using mongo (default NULL)}

\item{q}{parameter that controls the quantile of quickest URLs to be filtered. Default to 0.1 [0-1]}

\item{p}{parameter that controls the percentage of total shares to be reached. Default to 0.5 [0-1]}

\item{clean_urls}{clean up unnecessary url paramters and malformed urls, and keep just the URLs included in the original data set (default FALSE)}

\item{mongo_cluster}{logical: set to TRUE if you are using a MongoDB cluster instead of standalone instance (default FALSE)}

\item{keep_ourl_only}{restrict the analysis to ct shares links matching the original URLs (default=FALSE)}
}
\value{
a list containing two objects: summary statistics of q\% quickest second share performing URLs, and a time in seconds corresponding to the median time spent by these URLs to cumulate the p\% of their total shares
}
\description{
This function estimates a threshold in seconds that defines a coordinated link share. While it is common that multiple (pages/groups/account) entities share the same link, some tend to perform these actions in an unusually short period of time. Unusual is thus defined here as a function of the median co-share time difference. More specifically, the function ranks all co-shares by time-difference from first share and focuses on the behaviour of the quickest second share performing q\% (default 0.5) URLs. The value returned is the median time in seconds spent by these URLs to cumulate the p\% (default 0.1) of their total shares
}
\examples{
cord_int <- estimate_coord_interval(ct_shares.df = ct_shares.df, q=0.1, p=0.5, clean_urls=TRUE, keep_ourl_only=FALSE)
cord_int[[1]]
cord_int[[2]]

}
