% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_info.R
\name{get_bulk_raw_domain_info}
\alias{get_bulk_raw_domain_info}
\title{get_bulk_raw_domain_info}
\usage{
get_bulk_raw_domain_info(domain_names)
}
\arguments{
\item{domain_names}{a list of domain names to search}
}
\value{
A data.frame with info about the given domain names
        Note: all fields of the data.frame are set to NA except
        for raw_data
}
\description{
A function that returns raw the info about a list of domain
using direct calls to Whois servers
}
\examples{
  info <- get_free_bulk_domain_info("domain1.com")

}
