% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctshares_mongo.R
\name{get_ctshares_mongo}
\alias{get_ctshares_mongo}
\title{get_ctshares_mongo}
\usage{
get_ctshares_mongo(
  urls,
  url_column,
  date_column,
  platforms = "facebook,instagram",
  nmax = 1000,
  mongo_url = "",
  mongo_database = "",
  sleep_time = 30,
  mongo_cluster = TRUE,
  return_df = TRUE,
  get_history = FALSE,
  clean_urls = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{urls}{a dataframe with at least a column "url" containing the URLs, and a column "date" with their published date}

\item{url_column}{string: name of the column (placed inside quote marks) where the URLs are stored (defaults to "url")}

\item{date_column}{string: name of the column (placed inside quote marks) where the date of the URLs are stored (defaults to "date")}

\item{platforms}{string: default to "facebook,instagram". You can specify only "facebook" to search on Facebook, or only "instagram" to search on Instagram}

\item{nmax}{integer: max number of results for query (default 1000)}

\item{mongo_url}{string: address of the MongoDB server in standard URI Format (required, default empty string will raise an error)}

\item{mongo_database}{string: the name of the MongoDB used to host the collections. Name can be choosen by the user (required, default empty string  will raise an error)}

\item{sleep_time}{pause between queries to respect API rate limits. Default to 30 secs, it can be lowered or increased depending on the assigned \href{https://help.crowdtangle.com/en/articles/3443476-api-cheat-sheet}{API rate limit}.}

\item{mongo_cluster}{logical: set to TRUE if you are using a MongoDB cluster instead of standalone instance (default FALSE)}

\item{return_df}{logical: set to TRUE to return a dataframe of ct_shares (default FALSE)}

\item{get_history}{logical: set to TRUE to retrieve history from CrowdTangle API (default FALSE)}

\item{clean_urls}{logical: clean the URLs from tracking parameters (default FALSE)}

\item{verbose}{logical: show progress messages?}
}
\value{
a data.frame of posts that shared the URLs and a number of variables returned by the \href{https://github.com/CrowdTangle/API/wiki/Links}{CrowdTangle API links endpoint} and the original data set of news
}
\description{
A function to get the URLs shares from CrowdTangle
}
\details{
to start using the library you need to set the CrowdTangle API key.
  Open the environment variable file with file.edit("~/.Renviron"), write CROWDTANGLE_API_KEY = <YOUR_API_KEY>, save the file and restart your current R session to start using the CrowdTangle API
}
\examples{
  df <- get_ctshares_mongo(urls, url_column=“url”, date_column=“date”, platforms="facebook,instagram", nmax=100, sleep_time=30, clean_urls=FALSE, save_ctapi_output=FALSE)

}
