% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_info.R
\name{get_domain_info_from_api}
\alias{get_domain_info_from_api}
\title{get_domain_info_from_api}
\usage{
get_domain_info_from_api(hostname)
}
\arguments{
\item{hostname}{the domain name to search}
}
\value{
a data.frame with the response content or NULL
        if the server responded with a status that's
        not 200
}
\description{
A function that returns the domain info obtained
from the Whois API
}
