% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_info.R
\name{get_domains_from_summary}
\alias{get_domains_from_summary}
\title{get_domains_from_summary}
\usage{
get_domains_from_summary(
  summary,
  threshold = 0.8,
  limit = 5,
  excludes = "google.com, facebook.com, youtube.com",
  history = FALSE
)
}
\arguments{
\item{summary}{A data.frame produced by get_component_summary or get_cluster_summary}

\item{threshold}{A value between 0 and 1 used to filter the summary by
their gini.parent_domain}

\item{limit}{Limit the number of domains searched for each component}

\item{excludes}{A comma separated string of domain names to exclude from search}

\item{history}{If set to true returns the history data using the Whois API,
otherwise it returns the raw data}
}
\value{
A data.frame with info about the given domains and associated
        component from summary
}
\description{
A function that returns domains info from a summary data.frame
}
