% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_info.R
\name{get_empty_result_data}
\alias{get_empty_result_data}
\title{get_empty_result_data}
\usage{
get_empty_result_data()
}
\value{
A data.frame
}
\description{
A function that returns a empty return data.frame
}
