% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_info.R
\name{get_free_bulk_domain_info}
\alias{get_free_bulk_domain_info}
\title{get_free_bulk_domain_info}
\usage{
get_free_bulk_domain_info(domain_names)
}
\arguments{
\item{domain_names}{a list of domain names to search}
}
\value{
A data.frame with info about the given domain names
}
\description{
A function that returns all the info about a list of domain
using the Rwhois package
}
\examples{
  info <- get_free_bulk_domain_info("domain1.com")

}
