% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_info.R
\name{get_na_result_data}
\alias{get_na_result_data}
\title{get_na_result_data}
\usage{
get_na_result_data()
}
\value{
A data.frame
}
\description{
A function that returns a return data.frame with all NA values
}
