% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outputs_mongo.R
\name{get_outputs_mongo}
\alias{get_outputs_mongo}
\title{get_outputs_mongo}
\usage{
get_outputs_mongo(
  coord_shares_output,
  ct_shares_marked.df = TRUE,
  mongo_url = NULL,
  mongo_collection = NULL,
  mongo_cluster = FALSE,
  highly_connected_g = TRUE,
  highly_connected_coordinated_entities = TRUE,
  component_summary = FALSE,
  cluster_summary = FALSE,
  top_coord_urls = FALSE,
  top_coord_shares = FALSE,
  gdrive_folder_id = NULL
)
}
\arguments{
\item{coord_shares_output}{the list from the get_coord_shares function}

\item{ct_shares_marked.df}{a dataframe that contains all the retrieved social media shares plus an extra boolean variable (is_coordinated) that identify if the shares was coordinated (default TRUE)}

\item{mongo_url}{string: address of the MongoDB server in standard URI Format. Set to NULL to avoid using mongo (default NULL)}

\item{mongo_collection}{string: name of the MongoDB collection where the shares have been saved. Set to NULL to avoid using mongo (default NULL)}

\item{mongo_cluster}{logical: set to TRUE if you are using a MongoDB cluster instead of standalone instance (default FALSE)}

\item{highly_connected_g}{graph of coordinated networks of pages/groups/accounts (default TRUE)}

\item{highly_connected_coordinated_entities}{a dataframe that lists coordinated entities and corresponding component (default TRUE)}

\item{component_summary}{the result of CooRnet::get_component_summary (default FALSE)}

\item{cluster_summary}{the result of CooRnet::get_component_summary (default FALSE)}

\item{top_coord_urls}{the result of CooRnet::get_top_coord_urls (default FALSE)}

\item{top_coord_shares}{the result of CooRnet::get_top_coord_shares (default FALSE)}

\item{gdrive_folder_id}{a Google Drive folder id where to uploads selected files (output.rds, highly_connected_g.graphml, highly_connected_coordinated_entities.csv, component_summary.csv, top_coord_urls.csv, top_coord_shares) files (default NULL)}
}
\description{
A function to get the outputs of the get_coord_shares function
}
\examples{
get_outputs_mongo(coord_shares_output=output)

get_outputs_mongo(coord_shares_output=output, gdrive_folder_id="0BwwA4oUTeiV1TGRPeTVjaWRDY1E")

}
