% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_info.R
\name{get_raw_domain_info}
\alias{get_raw_domain_info}
\title{get_raw_domain_info}
\usage{
get_raw_domain_info(hostname, server)
}
\arguments{
\item{hostname}{the domain name to search}

\item{server}{the server address to use for the search}
}
\value{
A string with the raw domain info
}
\description{
A function that returns info about a domain in a row format
}
