% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_info.R
\name{get_single_domain_info}
\alias{get_single_domain_info}
\title{get_single_domain_info}
\usage{
get_single_domain_info(name)
}
\arguments{
\item{name}{the name of the domain to search}
}
\value{
A data.frame with info about the given domain name
}
\description{
A function that returns all the info about a single domain
}
\examples{
 info <- get_single_domain_info(name="domain1.com")

}
