% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_coord_urls.R
\name{get_top_coord_urls}
\alias{get_top_coord_urls}
\title{get_top_coord_urls}
\usage{
get_top_coord_urls(output, order_by = "engagement", component = TRUE, top = 10)
}
\arguments{
\item{output}{the output list resulting from the function get_coord_shares}

\item{order_by}{name of the column used to order the top news. Default to "engagement". Other possible values are: "statistics.actual.likeCount", "statistics.actual.shareCount", "statistics.actual.commentCount", "statistics.actual.loveCount", "statistics.actual.wowCount", "statistics.actual.hahaCount", "statistics.actual.sadCount","statistics.actual.angryCount"}

\item{component}{return the top URLs by network component (TRUE, default) or just the top URLs (FALSE)}

\item{top}{number of the top URLs to be retrieved}
}
\value{
A data frame (grouped_df) containing the top URLs shared in a coordinated way by the highly coordinated entities, with shares and engagement statistics, list of entities and components that shared the link
}
\description{
A function to get the top n URLs shared in a coordinated way
}
\examples{
# get the top ten URLs shared in a coordinated way by each network component, by engagement
df <- get_top_coord_urls(output, order_by = "engagement", component = TRUE, top=10)

# get the top ten URLs shared in a coordinated way, by engagement
df <- get_top_news(output, order_by = "engagement", top=10)

}
