% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urls_from_ct_histdata.R
\name{get_urls_from_ct_histdata}
\alias{get_urls_from_ct_histdata}
\title{get_urls_from_ct_histdata}
\usage{
get_urls_from_ct_histdata(ct_histdata_csv = NULL)
}
\arguments{
\item{ct_histdata_csv}{a local or remote link to a CSV created by the CrowdTangle Historical Data}
}
\value{
A data.frame with a unique list of URLs with respective first seen date
}
\description{
Given a CSV created by the CrowdTangle Historical Data feature (preferbly filtered for link type posts), this function extract a unique list of URLs with a first shared date
}
\examples{
urls <- get_urls_from_ct_histdata(ct_histdata_csv="mylocaldata.csv")

urls <- get_urls_from_ct_histdata(ct_histdata_csv="https://ct-download.s3.us-west-2.amazonaws.com/...")

# Use the new urls dataset to call get_ct_shares function
ct_shares.dt <- get_ctshares(urls, url_column="url", date_column="date", save_ctapi_output=TRUE)

}
