% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_info.R
\name{names_to_df}
\alias{names_to_df}
\title{names_to_df}
\usage{
names_to_df(x, FUN)
}
\arguments{
\item{x}{A vector of domain names}

\item{FUN}{A function applied to each domain name
Note: the function must return a data.frame}
}
\value{
A data.frame
}
\description{
A function that takes a vector of domain names and apply
a given FUN to each of them returning a data.frame
}
