% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_link_endpoint.R
\name{query_link_enpoint}
\alias{query_link_enpoint}
\title{query_link_enpoint}
\usage{
query_link_enpoint(query.string, sleep_time = 10)
}
\arguments{
\item{query.string}{a well formed query string for the link endpoint}

\item{sleep_time}{}
}
\value{
a data.frame of posts matching the given URL
}
\description{
A wrapper for CrowdTangle API Links Endpoint. Returns a dataframe of posts matching the given URL
}
\details{
to start using the library you need to set the CrowdTangle API key.
  Open the environment variable file with file.edit("~/.Renviron"), write CROWDTANGLE_API_KEY = <YOUR_API_KEY>, save the file and restart your current R session to start using the CrowdTangle API
}
