% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_info.R
\name{domains_info}
\alias{domains_info}
\title{domains_info}
\usage{
domains_info(
  domains,
  excludes = "google.com, facebook.com, youtube.com",
  history = FALSE
)
}
\arguments{
\item{domains}{A comma separated string of domain names}

\item{excludes}{A comma separated string of domain names to exclude from search}

\item{history}{If set to true returns the history data using the Whois API,
otherwise it returns the raw data}
}
\value{
A data.frame with info about the given domains
}
\description{
A function that returns Whois data about a given list of domain names
}
\details{
to start using the library you need to set the Whois API key.
  Open the environment variable file with file.edit("~/.Renviron"), write WHOIS_API_KEY = <YOUR_API_KEY>, save the file and restart your current R session to start using the Whois API
}
\examples{
 info <- domains_info(
   domains = "domain1.com, domain2.com, domain3.com",
   excludes="domain2.com",
   history=TRUE)

 info2 <- domains_info(
   domains = "domain1.com, domain2.com, domain3.com",
   excludes="domain2.com")
}
