% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_info.R
\name{domains_info}
\alias{domains_info}
\title{domains_info}
\usage{
domains_info(domains)
}
\arguments{
\item{domains}{A vector of domain names}
}
\value{
A data.frame with info about the given domains
}
\description{
A function that returns Whois data about a given list domain names
}
\examples{
 info <- domains_info(domains = c("google.com"))

}
