% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_component_summary.R
\name{get_component_summary}
\alias{get_component_summary}
\title{get_component_summary}
\usage{
get_component_summary(output, labels = FALSE)
}
\arguments{
\item{output}{A list output resulting from the function get_coord_shares.}

\item{labels}{Automatically generate a component label using account titles and descriptions. This relies on OpenAI APIs and expects the API Bearer in the OPENAI_API_KEY environment variable.}
}
\value{
A data frame that summarizes data for each coordinated component, including:
- The average number of subscribers of entities within a component.
- The proportion of coordinated shares over the total shares (coorshare_ratio).
- The average coordinated score (avg_cooRscore), which measures the dispersion (gini) in the distribution of domains that are coordinatedly shared by the component (0-1). Higher values correspond to higher concentration (fewer different domains linked).
- The top coordinatedly shared domains (ranked by the number of shares) and the total number of coordinatedly shared domains.
If the NewsGuard API is provided, this function also returns an estimate of the trustworthiness of the domains used by the component. If the label parameter is set to TRUE and an OpenAI token is provided, the function also returns an automatically generated label for each component.
}
\description{
A function to obtain a summary of data by coordinated component.
}
\details{
The Gini values are computed using the Gini coefficient on the proportions of unique domains each component shared. The Gini coefficient is a measure of the degree of concentration (inequality) of a variable in a distribution.
It ranges between 0 and 1: the more equal the distribution, the lower its Gini index. When a component shared just one domain, the value of the variable is set to 1. It is calculated separately for full_domains (e.g. www.foxnews.com, video.foxnews.com) and parent domains (foxnews.com).

The avg_cooRscore is a measure of component coordination. Higher values imply higher coordination.
Its value is calculated by dividing, for each entity in a coordinated network, its \code{\link[igraph]{strength}} by its \code{\link[igraph]{degree}}, and then calculating the average by component of these values.

The coorshare_ratio.avg is an additional measure of component coordination ranging from 0 (no shares coordinated) to 1 (all shares coordinated).
}
\examples{
# Build the component summary
component_summary <- get_component_summary(output, label=TRUE)

# component the components rowwise mutate
components <- hclust(dist(component_summary[, 2:4]))
plot(components)

}
