% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_component_summary.R
\name{get_component_summary}
\alias{get_component_summary}
\title{get_component_summary}
\usage{
get_component_summary(output)
}
\arguments{
\item{output}{the output list resulting from the function get_coord_shares}
}
\value{
A data frame containing summary data by each coordinated component:
the proportion of coordinated shares over the total shares (coorshare_ratio), the average coordinated score (avg_cooRscore), and a measure (gini) of how much the coordinated shares concern a more or less large number of domains
}
\description{
A function to get summary data by coordinated component
}
\details{
The gini values is computed by using the Gini coefficient on the proportions of unique domains each component shared.
The Gini coefficient is a measure of the degree of concentration (inequality) of a variable in a distribution.
It ranges between 0 and 1: the more nearly equal the distribution, the lower its Gini index. When a component shared just one domain, the value of the variable is set to 1.

The coorshare_ratio is a measure of how much coordinated is a component.
Its value is calculated by dividing, for each entity in a coordinated netwok, its \code{\link[igraph]{strength}} by its \code{\link[igraph]{degree}}, and then calculating the average by component of these values.
}
\examples{
  # get the top ten posts containing URLs shared by each network component and by engagement
  component_summary <- get_component_summary(output)

  # clustering the components
  clusters <- hclust(dist(component_summary[, 2:4]))
  plot(clusters)

}
