% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outputs.R
\name{get_outputs}
\alias{get_outputs}
\title{get_outputs}
\usage{
get_outputs(
  coord_shares_output,
  ct_shares_marked.df = TRUE,
  highly_connected_g = TRUE,
  highly_connected_coordinated_entities = TRUE
)
}
\arguments{
\item{coord_shares_output}{the list from the get_coord_shares function}

\item{ct_shares_marked.df}{a dataframe that contains all the retrieved social media shares plus an extra boolean variable (is_coordinated) that identify if the shares was coordinated (default TRUE)}

\item{highly_connected_g}{graph of coordinated networks of pages/groups/accounts (default TRUE)}

\item{highly_connected_coordinated_entities}{a dataframe that lists coordinated entities and corresponding component (default TRUE)}
}
\description{
A function to get the outputs of the get_coord_shares function
}
\examples{
get_outputs(get_coord_shares_output)

}
